Moved structs into respective paths
